requirements.txt
gradio
openai
anthropic
beautifulsoup4
requests
python-dotenv
